+++ aws-sdk-resources/lib/aws/resource/documenter.rb
@@ -11,6 +11,7 @@ module Aws
       autoload :OperationDocumenter, 'aws/resource/documenter/operation_documenter'
       autoload :ReferenceOperationDocumenter, 'aws/resource/documenter/reference_operation_documenter'
       autoload :ResourceOperationDocumenter, 'aws/resource/documenter/resource_operation_documenter'
+      autoload :WaiterOperationDocumenter, 'aws/resource/documenter/waiter_operation_documenter'
 
     end
   end